Coupon/filters.py
import django_filters as filters
from django.db.models import Q
from unidecode import unidecode


class CouponFilters(filters.FilterSet):
    search = filters.CharFilter(method="search_filter")
    limit = filters.CharFilter(method="limit_filter")

    @staticmethod
    def search_filter(qs, name, value):
        qs = qs.filter(
            Q(code__icontains=value) |
            Q(percent__icontains=value) |
            Q(uses_number__icontains=value)
        ).distinct()
        return qs

    @staticmethod
    def limit_filter(qs, name, value):
        try:
            qs = qs.distinct()[:int(unidecode(value))]
        except:
            pass
        return qs
